requirements.txt
asgiref==3.6.0
Django==3.2.10
django-filter==21.1
djangorestframework==3.12.4
Pillow==9.4.0
psycopg2-binary==2.9.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
